requirements.txt
streamlit
faiss-cpu
numpy
sentence-transformers
ollama
pillow
python-dotenv